requirements.txt
beautifulsoup4==4.6.3
Django==2.0.7
django-admin-sortable2==0.6.19
django-cleanup==2.1.0
django-colorful==1.2
django-debug-toolbar==1.9.1
django-htmlmin==0.10.0
django-jet==1.0.7
django-js-asset==1.1.0
django-mptt==0.9.0
django-solo==1.1.3
django-widget-tweaks==1.4.2
django-wysiwyg-redactor==0.5.1
html5lib==1.0.1
Pillow==5.2.0
pytz==2018.5
six==1.11.0
sqlparse==0.2.4
webencodings==0.5.1
